Allow user to input course names and decide whether to continue adding courses
This commit adds functionality to allow the user to input course names interactively. The program prompts the user to enter the name of each course and provides an option to continue adding more courses or finish inputting. The entered course names are stored in the courseNames vector for later use in the CGPA calculation.